Убирает массивы с данными в отдельный модуль, заводит отдельный модуль под генерацию массива объектов, убирает заглушку с заголовка списка фотографий.
