+++ js/getPictures.js
@@ -1,4 +1,4 @@
-import {getPhotoArray} from './data.js';
+import {getPhotoArray} from './getPhotoArray.js';
 
 const picturesContainer = document.querySelector('.pictures');
 const pictureTemplate = document.querySelector('#picture')
@@ -19,6 +19,7 @@ const getPictures = () => {
 
   return picturesContainer.append(picturesFragment);
 };
-
+//Убираем заглушку с заголовка списка фотографий
+picturesContainer.querySelector('.pictures__title').classList.remove('visually-hidden');
 
 export {getPictures};
